requirements.txt
flask==2.3.3
pandas==2.2.2
numpy==1.26.4
plotly==5.24.1
werkzeug==3.0.3
flask-login==0.6.3
azure-identity==1.17.1
azure-keyvault-secrets==4.7.0
openpyxl==3.1.5
pdfplumber==0.11.4